+++ sdk/eventhubs/azure-messaging-eventhubs/src/test/java/com/azure/messaging/eventhubs/EventHubProducerAsyncClientIntegrationTest.java
@@ -3,32 +3,26 @@
 
 package com.azure.messaging.eventhubs;
 
-import com.azure.core.amqp.exception.AmqpException;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.messaging.eventhubs.models.CreateBatchOptions;
 import com.azure.messaging.eventhubs.models.SendOptions;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import reactor.test.StepVerifier;
 
-import java.time.Duration;
-import java.time.Instant;
 import java.util.Arrays;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
-import java.util.stream.IntStream;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
-public class EventHubProducerAsyncClientIntegrationTest extends IntegrationTestBase {
+class EventHubProducerAsyncClientIntegrationTest extends IntegrationTestBase {
     private static final String PARTITION_ID = "1";
 
     private EventHubProducerAsyncClient producer;
 
-    public EventHubProducerAsyncClientIntegrationTest() {
+    EventHubProducerAsyncClientIntegrationTest() {
         super(new ClientLogger(EventHubProducerAsyncClientIntegrationTest.class));
     }
 
@@ -49,7 +43,7 @@ public class EventHubProducerAsyncClientIntegrationTest extends IntegrationTestB
      * Verifies that we can create and send a message to an Event Hub partition.
      */
     @Test
-    public void sendMessageToPartition() {
+    void sendMessageToPartition() {
         // Arrange
         final SendOptions sendOptions = new SendOptions().setPartitionId(PARTITION_ID);
         final List<EventData> events = Arrays.asList(
@@ -67,7 +61,7 @@ public class EventHubProducerAsyncClientIntegrationTest extends IntegrationTestB
      * the service distribute the events.
      */
     @Test
-    public void sendMessage() {
+    void sendMessage() {
         // Arrange
         final List<EventData> events = Arrays.asList(
             new EventData("Event 1".getBytes(UTF_8)),
@@ -83,7 +77,7 @@ public class EventHubProducerAsyncClientIntegrationTest extends IntegrationTestB
      * Verifies we can create an {@link EventDataBatch} and send it using our EventHubProducer.
      */
     @Test
-    public void sendBatch() {
+    void sendBatch() {
         // Arrange
         final List<EventData> events = Arrays.asList(
             new EventData("Event 1".getBytes(UTF_8)),
@@ -107,7 +101,7 @@ public class EventHubProducerAsyncClientIntegrationTest extends IntegrationTestB
      * Verifies we can create an {@link EventDataBatch} with a partition key and send it using our EventHubProducer.
      */
     @Test
-    public void sendBatchWithPartitionKey() {
+    void sendBatchWithPartitionKey() {
         // Arrange
         final List<EventData> events = Arrays.asList(
             new EventData("Event 1".getBytes(UTF_8)),
@@ -135,7 +129,7 @@ public class EventHubProducerAsyncClientIntegrationTest extends IntegrationTestB
      * Verify that we can send to multiple partitions, round-robin, and with a partition key, using the same producer.
      */
     @Test
-    public void sendEventsWithKeyAndPartition() {
+    void sendEventsWithKeyAndPartition() {
         // Arrange
         final List<EventData> events = Arrays.asList(
             new EventData("Event 1".getBytes(UTF_8)),
@@ -155,7 +149,7 @@ public class EventHubProducerAsyncClientIntegrationTest extends IntegrationTestB
     }
 
     @Test
-    public void sendAllPartitions() {
+    void sendAllPartitions() {
         final List<String> partitionIds = producer.getPartitionIds().collectList().block(TIMEOUT);
 
         Assertions.assertNotNull(partitionIds);
@@ -177,7 +171,7 @@ public class EventHubProducerAsyncClientIntegrationTest extends IntegrationTestB
      * Sending with credentials.
      */
     @Test
-    public void sendWithCredentials() {
+    void sendWithCredentials() {
         // Arrange
         final EventData event = new EventData("body");
         final SendOptions options = new SendOptions().setPartitionId(PARTITION_ID);
@@ -200,37 +194,5 @@ public class EventHubProducerAsyncClientIntegrationTest extends IntegrationTestB
         } finally {
             dispose(client);
         }
-    }
-
-    @Disabled("Testing long running operations and disconnections.")
-    @Test
-    void worksAfterReconnection() throws InterruptedException {
-        Flux.interval(Duration.ofSeconds(5))
-            .flatMap(position -> producer.createBatch().flatMap(batch -> {
-                IntStream.range(0, 3).mapToObj(number -> new EventData("Position" + position + ": " + number))
-                    .forEach(event -> {
-                        if (!batch.tryAdd(event)) {
-                            logger.error("Could not add event. Size: {}. Max: {}. Content: {}",
-                                batch.getSizeInBytes(), batch.getMaxSizeInBytes(), event.getBodyAsString());
-                        }
-                    });
-
-                return producer.send(batch).thenReturn(Instant.now());
-            }))
-            .onErrorContinue(error -> error instanceof AmqpException && ((AmqpException) error).isTransient(),
-                (error, value) -> {
-                    System.out.println("Retries were exhausted. No logger retrying operation. " + error.getMessage());
-                })
-            .subscribe(instant -> {
-                System.out.println("Sent batch at: " + instant);
-            }, error -> {
-                    logger.error("Error sending batch: ", error);
-                }, () -> {
-                    logger.info("Complete.");
-                });
-
-        System.out.println("Sleeping while performing work.");
-        TimeUnit.MINUTES.sleep(30);
-        System.out.println("Complete.");
     }
 }