+++ sdk/eventhubs/azure-messaging-eventhubs/src/test/java/com/azure/messaging/eventhubs/implementation/ReactorConnectionIntegrationTest.java
@@ -18,7 +18,6 @@ import com.azure.core.credential.TokenCredential;
 import com.azure.core.util.CoreUtils;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.messaging.eventhubs.IntegrationTestBase;
-import java.util.Map;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -29,6 +28,7 @@ import reactor.test.StepVerifier;
 
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
+import java.util.Map;
 
 import static com.azure.core.amqp.implementation.CbsAuthorizationType.SHARED_ACCESS_SIGNATURE;
 
@@ -76,7 +76,7 @@ public class ReactorConnectionIntegrationTest extends IntegrationTestBase {
     @Override
     protected void afterTest() {
         if (connection != null) {
-            connection.close();
+            connection.dispose();
         }
     }
 
@@ -96,7 +96,7 @@ public class ReactorConnectionIntegrationTest extends IntegrationTestBase {
             getConnectionStringProperties().getEndpoint().getHost(),
             ClientConstants.AZURE_ACTIVE_DIRECTORY_SCOPE);
 
-        final String tokenAudience = provider.getResourceString(getConnectionStringProperties().getEntityPath());
+        final String tokenAudience = provider.getScopesFromResource(getConnectionStringProperties().getEntityPath());
 
         // Act & Assert
         StepVerifier.create(connection.getClaimsBasedSecurityNode().flatMap(node -> node.authorize(tokenAudience, tokenAudience)))