+++ sdk/storage/azure-storage-blob-batch/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
 
+## 12.3.0 (2020-01-16)
+This package's
+[documentation](https://github.com/Azure/azure-sdk-for-java/blob/azure-storage-blob-batch_12.3.0/sdk/storage/azure-storage-blob-batch/README.md)
+and
+[samples](https://github.com/Azure/azure-sdk-for-java/blob/azure-storage-blob-batch_12.3.0/sdk/storage/azure-storage-blob-batch/src/samples/java/com/azure/storage/blob/batch)
+
+- Upgraded to version 12.3.0 of Azure Storage Blob.
+
 ## 12.2.0 (2020-01-08)
 This package's
 [documentation](https://github.com/Azure/azure-sdk-for-java/blob/azure-storage-blob-batch_12.2.0/sdk/storage/azure-storage-blob-batch/README.md)
@@ -7,6 +15,7 @@ and
 [samples](https://github.com/Azure/azure-sdk-for-java/blob/azure-storage-blob-batch_12.2.0/sdk/storage/azure-storage-blob-batch/src/samples/java/com/azure/storage/blob/batch)
 
 - Upgraded to version 12.2.0 of Azure Storage Blob.
+- Updated batch request creation to prevent race condition which could result in a dropped operation.
 
 ## 12.1.0 (2019-12-04)
 This package's